fix: excluding test/ from phpstan makes it impossible to bump to v2 because of unused trait (ES-99)
