+++ composer.json
@@ -33,7 +33,7 @@
         "squizlabs/php_codesniffer": "^3",
         "dealerdirect/phpcodesniffer-composer-installer": "^1",
         "phpcompatibility/php-compatibility": "^9",
-        "phpstan/phpstan": "^1"
+        "phpstan/phpstan": "^2"
     },
     "autoload-dev": {
         "psr-4": {
