+++ phpstan.neon
@@ -1,8 +1,6 @@
 parameters:
     level: 9
     paths:
+        - example
         - src
         - test
-    excludePaths:
-        analyse:
-            - test
