+++ test/DynamicAccessorTest.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace EntireStudio\DynamicAccessors\Test;
 
-use EntireStudio\DynamicAccessors\{DynamicAccessors, Get, Set};
+use EntireStudio\DynamicAccessors\Test\Stub\{
+    Cut,
+    CutWithConstructorAndPrivateProperties
+};
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
 
@@ -16,16 +19,7 @@ class DynamicAccessorTest extends TestCase
     #[TestDox('Accessors declared on properties')]
     public function testAccessorsOnProperties(): void
     {
-        $cut = new class
-        {
-            use DynamicAccessors;
-
-            #[Set, Get]
-            protected string $firstName = 'John';
-
-            #[Set('setLastName'), Get('getLastName')]
-            private string $lastName = 'Doe';
-        };
+        $cut = new Cut();
 
         $this->assertEquals($this->firstName, $cut->firstName());
         $cut->firstName('NOT ' . $this->firstName);
@@ -39,18 +33,10 @@ class DynamicAccessorTest extends TestCase
     #[TestDox('Accessors declared on constructor arguments')]
     public function testAccessorsOnConstructorArguments(): void
     {
-        $cut = new class ($this->firstName, $this->lastName)
-        {
-            use DynamicAccessors;
-
-            public function __construct(
-                #[Set, Get]
-                private string $firstName,
-                #[Set('setLastName'), Get('getLastName')]
-                private string $lastName,
-            ) {
-            }
-        };
+        $cut = new CutWithConstructorAndPrivateProperties(
+            $this->firstName,
+            $this->lastName
+        );
 
         $this->assertEquals($this->firstName, $cut->firstName());
         $cut->firstName('NOT ' . $this->firstName);
